requirements.txt
fastapi
uvicorn #servidor local
pdfplumber
python-multipart
aiofiles
groq
python-dotenv
